pyproject.toml
[project]
name = "MultimediaSpider"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "weixiaolong", email = "350944197@qq.com"},
]
dependencies = [
    "aiofiles==24.1.0",
    "aiomysql==0.2.0",
    "parsel==1.6.0",
    "Pillow==10.4.0",
    "httpx==0.24.0",
    "tenacity==8.2.2",
    "pydantic==2.5.2",
    "redis==4.6.0",
    "openpyxl",
    "pandas",
    "cryptography",
    "aiohttp",
    "logto>=0.2.1",
    "sanic>=24.6.0",
    "tortoise-orm>=0.21.6",
    "sanic-session>=0.8.0",
    "PyExecJS>=1.5.1",
    "apscheduler>=3.10.4",
    "aerich>=0.7.2",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "sanic server.app --dev"
[tool.aerich]
tortoise_orm = "server.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
